restructure payment logic and gateways
